Update test case to reflect the new download path
With the updated download_all method, the images are downloaded to target_dir/images/dataset_name irrespective of number of datasets. This is more consistent behaviour.

Previously, the images were downloaded to target_dir/images if there was only one dataset in the project and target_dir/images/dataset_name if there were multiple datasets